chore(blueprint-cli): rename files for consistency
