+++ packages/utils/blueprint-cli/src/ast/parser/handle-declaration.ts
@@ -1,5 +1,6 @@
 import { extractJsdoc } from './extract-js-doc';
-import { convertToNode, Node } from './parser';
+import { Node } from './node';
+import { convertToNode } from './parse';
 
 const isOptional = (property): boolean => {
   return property.questionToken ? true : false;
