+++ packages/utils/blueprint-cli/src/ast/parser/walk.spec.ts
@@ -1,13 +1,14 @@
 import * as astStringArray from './asts/arrays/ast-array-string.json';
+import { Node, SupportedKind } from './node';
 
-import { extractProperties, Node, SupportedKind } from './parser';
+import { parse } from './parse';
 import { walk } from './walk';
 
 describe('AST walking', () => {
   describe('Depth first walking', () => {
     it('should have walk the elements on the AST', () => {
       const elements: Node[] = [];
-      for (const node of walk(extractProperties(JSON.stringify(astStringArray))[0])) {
+      for (const node of walk(parse(JSON.stringify(astStringArray))[0])) {
         elements.push(node);
       }
